+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 - Field `cors` added to the `http_server` input and output, for supporting CORS requests when custom servers are used.
 - Field `server_side_encryption` added to the `aws_s3` output.
 - Field `use_histogram_timing` added to the `prometheus` metrics exporter.
+- New duration string and back off field types added to plugin config builders.
 
 ### Changed